Revert "fix: 减少仪表板请求"
This reverts commit 950b25e04a1d40d31fa9ea6d22e5c71b92d6787f.
