+++ core/core-frontend/src/components/dashboard/DbToolbar.vue
@@ -22,26 +22,20 @@ import { ElIcon, ElMessage, ElMessageBox } from 'element-plus-secondary'
 import eventBus from '@/utils/eventBus'
 import { useEmbedded } from '@/store/modules/embedded'
 import { deepCopy } from '@/utils/utils'
-import {
-  nextTick,
-  reactive,
-  ref,
-  computed,
-  onBeforeUnmount,
-  onMounted,
-  defineAsyncComponent
-} from 'vue'
+import { nextTick, reactive, ref, computed, toRefs, onBeforeUnmount, onMounted } from 'vue'
 import { dvMainStoreWithOut } from '@/store/modules/data-visualization/dvMain'
 import { useAppStoreWithOut } from '@/store/modules/app'
 import { snapshotStoreWithOut } from '@/store/modules/data-visualization/snapshot'
 import { storeToRefs } from 'pinia'
 import Icon from '../icon-custom/src/Icon.vue'
 import ComponentGroup from '@/components/visualization/ComponentGroup.vue'
+import UserViewGroup from '@/custom-component/component-group/UserViewGroup.vue'
 import QueryGroup from '@/custom-component/component-group/QueryGroup.vue'
 import MediaGroup from '@/custom-component/component-group/MediaGroup.vue'
 import TextGroup from '@/custom-component/component-group/TextGroup.vue'
 import ComponentButton from '@/components/visualization/ComponentButton.vue'
 import ComponentButtonLabel from '@/components/visualization/ComponentButtonLabel.vue'
+import MultiplexingCanvas from '@/views/common/MultiplexingCanvas.vue'
 import { useI18n } from '@/hooks/web/useI18n'
 import { getPanelAllLinkageInfo, saveLinkage } from '@/api/visualization/linkage'
 import { queryVisualizationJumpInfo } from '@/api/visualization/linkJump'
@@ -55,10 +49,13 @@ import {
 import { useEmitt } from '@/hooks/web/useEmitt'
 import { copyStoreWithOut } from '@/store/modules/data-visualization/copy'
 import TabsGroup from '@/custom-component/component-group/TabsGroup.vue'
+import DeResourceGroupOpt from '@/views/common/DeResourceGroupOpt.vue'
+import OuterParamsSet from '@/components/visualization/OuterParamsSet.vue'
 import { XpackComponent } from '@/components/plugin'
 import DbMoreComGroup from '@/custom-component/component-group/DbMoreComGroup.vue'
 import { useCache } from '@/hooks/web/useCache'
 import DeFullscreen from '@/components/visualization/common/DeFullscreen.vue'
+import DeAppApply from '@/views/common/DeAppApply.vue'
 import { useUserStoreWithOut } from '@/store/modules/user'
 import { updatePublishStatus } from '@/api/visualization/dataVisualization'
 const { t } = useI18n()
@@ -67,7 +64,6 @@ const snapshotStore = snapshotStoreWithOut()
 const copyStore = copyStoreWithOut()
 const { styleChangeTimes, snapshotIndex } = storeToRefs(snapshotStore)
 const resourceAppOpt = ref(null)
-const resourceAppOptShow = ref(false)
 const {
   linkageSettingStatus,
   curLinkageView,
@@ -78,28 +74,10 @@ const {
   batchOptStatus,
   targetLinkageInfo,
   curBatchOptComponents,
-  appData
+  appData,
+  hiddenListStatus
 } = storeToRefs(dvMainStore)
 const dvModel = 'dashboard'
-const MultiplexingCanvasShow = ref(false)
-const MultiplexingCanvas = defineAsyncComponent(
-  () => import('@/views/common/MultiplexingCanvas.vue')
-)
-
-const OuterParamsSet = defineAsyncComponent(
-  () => import('@/components/visualization/OuterParamsSet.vue')
-)
-
-const UserViewGroup = defineAsyncComponent(
-  () => import('@/custom-component/component-group/UserViewGroup.vue')
-)
-
-const DeResourceGroupOpt = defineAsyncComponent(
-  () => import('@/views/common/DeResourceGroupOpt.vue')
-)
-
-const DeAppApply = defineAsyncComponent(() => import('@/views/common/DeAppApply.vue'))
-
 const multiplexingRef = ref(null)
 const fullScreeRef = ref(null)
 let nameEdit = ref(false)
@@ -110,14 +88,21 @@ const state = reactive({
   preBatchCanvasViewInfo: {}
 })
 const resourceGroupOpt = ref(null)
-const resourceGroupOptShow = ref(false)
 const outerParamsSetRef = ref(null)
-const outerParamsSetShow = ref(false)
 const { wsCache } = useCache('localStorage')
 const userStore = useUserStoreWithOut()
 const isIframe = computed(() => appStore.getIsIframe)
 const desktop = wsCache.get('app.desktop')
 const emits = defineEmits(['recoverToPublished'])
+
+const props = defineProps({
+  createType: {
+    type: String,
+    default: 'create'
+  }
+})
+
+const { createType } = toRefs(props)
 
 const editCanvasName = () => {
   nameEdit.value = true
@@ -253,16 +238,12 @@ const saveCanvasWithCheck = (withPublish = false, status?) => {
         },
         appData: appData.value
       }
-      resourceAppOptShow.value = true
       nextTick(() => {
         resourceAppOpt.value.init(params)
       })
     } else {
       const params = { name: dvInfo.value.name, leaf: true, id: dvInfo.value.pid || '0' }
-      resourceGroupOptShow.value = true
-      nextTick(() => {
-        resourceGroupOpt.value.optInit('leaf', params, 'newLeaf', true, { withPublish, status })
-      })
+      resourceGroupOpt.value.optInit('leaf', params, 'newLeaf', true, { withPublish, status })
       return
     }
   }
@@ -299,10 +280,7 @@ const saveResource = (checkParams?) => {
             () => {
               useEmitt().emitter.emit('refresh-dataset-selector')
               useEmitt().emitter.emit('calcData-all')
-              resourceAppOptShow.value = true
-              nextTick(() => {
-                resourceAppOpt.value.close()
-              })
+              resourceAppOpt.value.close()
               dvMainStore.setAppDataInfo(null)
               snapshotStore.resetSnapshot()
             }
@@ -373,10 +351,7 @@ const backHandler = (url: string) => {
 }
 
 const multiplexingCanvasOpen = () => {
-  MultiplexingCanvasShow.value = true
-  nextTick(() => {
-    multiplexingRef.value.dialogInit()
-  })
+  multiplexingRef.value.dialogInit()
 }
 onMounted(() => {
   eventBus.on('preview', previewInner)
@@ -476,10 +451,7 @@ const openOuterParamsSet = () => {
   }
   //设置需要先触发保存
   canvasSave(() => {
-    outerParamsSetShow.value = true
-    nextTick(() => {
-      outerParamsSetRef.value.optInit()
-    })
+    outerParamsSetRef.value.optInit()
   })
 }
 
@@ -551,7 +523,6 @@ const initOpenHandler = newWindow => {
     openHandler.value.invokeMethod(pm)
   }
 }
-const userGroupShow = ref(false)
 </script>
 
 <template>
@@ -613,13 +584,8 @@ const userGroupShow = ref(false)
             :icon-name="dvView"
             themes="light"
             :title="t('chart.datalist')"
-            @show="userGroupShow = true"
           >
-            <user-view-group
-              v-if="userGroupShow"
-              themes="light"
-              :dv-model="dvModel"
-            ></user-view-group>
+            <user-view-group themes="light" :dv-model="dvModel"></user-view-group>
           </component-group>
           <component-group
             :base-width="115"
@@ -868,14 +834,13 @@ const userGroupShow = ref(false)
       />
     </Teleport>
 
-    <multiplexing-canvas v-if="MultiplexingCanvasShow" ref="multiplexingRef"></multiplexing-canvas>
+    <multiplexing-canvas ref="multiplexingRef"></multiplexing-canvas>
     <de-resource-group-opt
       @finish="resourceOptFinish"
       cur-canvas-type="dashboard"
-      v-if="resourceGroupOptShow"
       ref="resourceGroupOpt"
     />
-    <outer-params-set v-if="outerParamsSetShow" ref="outerParamsSetRef"> </outer-params-set>
+    <outer-params-set ref="outerParamsSetRef"> </outer-params-set>
   </div>
   <de-fullscreen show-position="edit" ref="fullScreeRef"></de-fullscreen>
   <XpackComponent ref="openHandler" jsname="L2NvbXBvbmVudC9lbWJlZGRlZC1pZnJhbWUvT3BlbkhhbmRsZXI=" />
@@ -885,7 +850,6 @@ const userGroupShow = ref(false)
     :dv-info="dvInfo"
     :canvas-view-info="canvasViewInfo"
     cur-canvas-type="dashboard"
-    v-if="resourceAppOptShow"
     @saveAppCanvas="saveCanvasWithCheck"
   ></de-app-apply>
 </template>
