+++ core/core-frontend/src/components/visualization/ComponentGroup.vue
@@ -25,10 +25,6 @@ const props = defineProps({
 })
 
 const { title, iconName, baseWidth, themes } = toRefs(props)
-const emits = defineEmits(['show'])
-const handleVisibleChange = () => {
-  emits('show')
-}
 </script>
 
 <template>
@@ -37,7 +33,6 @@ const handleVisibleChange = () => {
     :width="baseWidth"
     trigger="click"
     :show-arrow="false"
-    @show="handleVisibleChange"
     :popper-class="'custom-popover-' + themes"
   >
     <template #reference>
