+++ core/core-frontend/src/custom-component/de-tabs/Component.vue
@@ -154,8 +154,7 @@ import {
   reactive,
   ref,
   toRefs,
-  watch,
-  defineAsyncComponent
+  watch
 } from 'vue'
 import DeCanvas from '@/views/canvas/DeCanvas.vue'
 import { dvMainStoreWithOut } from '@/store/modules/data-visualization/dvMain'
@@ -169,6 +168,7 @@ import {
   isDashboard
 } from '@/utils/canvasUtils'
 import DeCustomTab from '@/custom-component/de-tabs/DeCustomTab.vue'
+import DePreview from '@/components/data-visualization/canvas/DePreview.vue'
 import { getPanelAllLinkageInfo } from '@/api/visualization/linkage'
 import { dataVTabComponentAdd, groupSizeStyleAdaptor } from '@/utils/style'
 import { deepCopyTabItemHelper } from '@/store/modules/data-visualization/copy'
@@ -242,9 +242,6 @@ const {
   searchCount
 } = toRefs(props)
 
-const DePreview = defineAsyncComponent(
-  () => import('@/components/data-visualization/canvas/DePreview.vue')
-)
 const titleBackgroundActiveSvgInner = computed(() => {
   return element.value.titleBackground.active.innerImage.replace('board/', '').replace('.svg', '')
 })
