+++ core/core-frontend/src/pages/index/App.vue
@@ -1,19 +1,13 @@
 <script setup lang="ts">
-import { ref, defineAsyncComponent, nextTick } from 'vue'
+import { ref } from 'vue'
 import { useEmitt } from '@/hooks/web/useEmitt'
 import configGlobal from '@/components/config-global/src/ConfigGlobal.vue'
 import { useRoute } from 'vue-router'
+import ExportExcel from '@/views/visualized/data/dataset/ExportExcel.vue'
 const route = useRoute()
 const exportExcelRef = ref()
-const exportExcelRefShow = ref(false)
-const ExportExcel = defineAsyncComponent(
-  () => import('@/views/visualized/data/dataset/ExportExcel.vue')
-)
 const exportExcelCenter = params => {
-  exportExcelRefShow.value = true
-  nextTick(() => {
-    exportExcelRef.value.init(params)
-  })
+  exportExcelRef.value.init(params)
 }
 useEmitt({
   name: 'data-export-center',
@@ -23,6 +17,6 @@ useEmitt({
 <template>
   <config-global>
     <router-view :key="route.path" />
-    <ExportExcel v-if="exportExcelRefShow" ref="exportExcelRef"></ExportExcel>
+    <ExportExcel ref="exportExcelRef"></ExportExcel>
   </config-global>
 </template>
