+++ core/core-frontend/src/views/data-visualization/indexV3.vue
@@ -10,7 +10,7 @@ import { useAppStoreWithOut } from '@/store/modules/app'
 import { storeToRefs } from 'pinia'
 import DvToolbar from '../../components/data-visualization/DvToolbar.vue'
 import eventBus from '../../utils/eventBus'
-import { findComponentAttr } from '@/utils/componentsAttr'
+import { findComponentAttr } from '../../utils/components'
 import DvSidebar from '../../components/visualization/DvSidebar.vue'
 import router from '@/router'
 import Editor from '@/views/chart/components/editor/index.vue'
